feat(frontend): define authenticated header component ss-182
